NIFI-10659: Extracted DBCP common code to nifi-dbcp-base module for use in Snowflake NAR
