+++ nifi-nar-bundles/nifi-snowflake-bundle/nifi-snowflake-services-nar/pom.xml
@@ -33,7 +33,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-dbcp-service-nar</artifactId>
+            <artifactId>nifi-standard-services-api-nar</artifactId>
             <version>1.19.0-SNAPSHOT</version>
             <type>nar</type>
         </dependency>
